Enhanced the daemon to use encrypted connections over SSL/TLS.
The new fmhub is not a executable but a library.  The library uses a
private library to die().  So this patch replaces the local die()
function in libutil to libdie in fsyscall.

This patch also changes the directory structure.  A settings directory
will not include only a configuration file, but also an SSL certificate
and a private key.